move NSQLookup back to a interface so jackson json parser is not forced on everyone
